+++ kitty/layout/splits.py
@@ -12,7 +12,7 @@ from kitty.window_list import WindowGroup, WindowList
 
 from .base import (
     Layout, LayoutOpts, NeighborsMap, blank_rects_for_window, lgd,
-    window_geometry_from_layouts
+    window_geometry_from_layouts, BorderLine
 )
 
 
@@ -417,14 +417,14 @@ class Splits(Layout):
             pair.bias = 0.5
         return True
 
-    def window_independent_borders(self, all_windows: WindowList) -> Generator[Edges, None, None]:
+    def window_independent_borders(self, all_windows: WindowList) -> Generator[BorderLine, None, None]:
         groups = tuple(all_windows.iter_all_layoutable_groups())
         window_count = len(groups)
         if not lgd.draw_minimal_borders or window_count < 2:
             return
         for pair in self.pairs_root.self_and_descendants():
             if pair.between_border is not None:
-                yield pair.between_border
+                yield BorderLine(pair.between_border)
 
     def neighbors_for_window(self, window: WindowType, all_windows: WindowList) -> NeighborsMap:
         wg = all_windows.group_for_window(window)